pyproject.toml
[project]
name = "cv-adapter"
version = "0.1.0"
description = "API para adaptar CVs a ofertas laborales usando IA"
authors = [
    {name = "Diego", email = "diego@example.com"}
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "langgraph-cli[inmem]>=0.4.11",
    "redis>=5.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.24.0"
]
